requirements.txt
nltk
hjsystemtools
langdetect
unidecode
guess_language-spirit
pyenchant